Add checksum to compressed index.
Summary:
 - Add CRC32 checksum for memsector data.

closes #29.

Test Plan:
mws-index -I data/zbl -o data/zbl-compressed
mwsd-load -I data/zbl-compressed

Reviewers: radu.hambasan

Reviewed By: radu.hambasan

Differential Revision: http://phab.code4fun.de/D62

+++ src/mws/index/memsector_allocator.h
@@ -66,7 +66,7 @@ memsector_off_t memsector_alloc_get_curr_off(
 }
 
 static inline
-uint32_t memsector_size_inuse(const memsector_alloc_header_t* alloc) {
+uint64_t memsector_size_inuse(const memsector_alloc_header_t* alloc) {
     return alloc->curr_offset;
 }
